challenge-weather-data/weather_data_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]
vertexai>=1.0.0
google-cloud-bigquery
google-cloud-storage
google-api-python-client
google-auth-httplib2
